requirements.txt
requests
pyyaml
argparse
colorlog
Pillow
opencv-python
